+++ src/web/assets/garnish/src/Garnish.js
@@ -41,6 +41,7 @@ Garnish.ltr = !Garnish.rtl;
 
 Garnish = $.extend(Garnish, {
   $scrollContainer: Garnish.$win,
+  resizeEventsMuted: false,
 
   // Key code constants
   BACKSPACE_KEY: 8,
@@ -908,6 +909,13 @@ Garnish = $.extend(Garnish, {
       }
     }
   },
+
+  muteResizeEvents: function (callback) {
+    const resizeEventsMuted = Garnish.resizeEventsMuted;
+    Garnish.resizeEventsMuted = true;
+    callback();
+    Garnish.resizeEventsMuted = resizeEventsMuted;
+  },
 });
 
 Object.assign(Garnish, {
@@ -957,7 +965,9 @@ function getErd() {
 }
 
 function triggerResizeEvent(elem) {
-  $(elem).trigger('resize');
+  if (!Garnish.resizeEventsMuted) {
+    $(elem).trigger('resize');
+  }
 }
 
 // Work them into jQuery's event system